+++ coderd/organizations_test.go
@@ -37,4 +37,3 @@ func TestOrganizationByUserAndName(t *testing.T) {
 		require.NoError(t, err)
 	})
 }
-
