+++ enterprise/cli/templatecreate_test.go
@@ -140,7 +140,10 @@ func TestTemplateCreate(t *testing.T) {
 		t.Parallel()
 
 		dv := coderdtest.DeploymentValues(t)
-		dv.Experiments = []string{string(codersdk.ExperimentCustomRoles)}
+		dv.Experiments = []string{
+			string(codersdk.ExperimentCustomRoles),
+			string(codersdk.ExperimentMultiOrganization),
+		}
 		ownerClient, _ := coderdenttest.New(t, &coderdenttest.Options{
 			Options: &coderdtest.Options{
 				DeploymentValues: dv,
@@ -152,6 +155,7 @@ func TestTemplateCreate(t *testing.T) {
 					codersdk.FeatureAccessControl:              1,
 					codersdk.FeatureCustomRoles:                1,
 					codersdk.FeatureExternalProvisionerDaemons: 1,
+					codersdk.FeatureMultipleOrganizations:      1,
 				},
 			},
 		})
